[MOL-12849] Split into specific helper files and refactored types
